pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["teddy <quddnr135@naver.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
matplotlib = "^3.8.4"
langchain = "^0.3.13"
langgraph = "^0.2.60"
langchain-openai = "^0.2.14"
jupyter = "^1.1.1"
notebook = "^7.3.1"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
scikit-learn = "^1.3.0"
tavily-python = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
